test(e2e): split app.e2e-spec.ts into multiple logical test files for better structure
