--- apps/backend/test/app.e2e-spec.ts
@@ -1,913 +1,0 @@
-import { Test, TestingModule } from "@nestjs/testing";
-import { INestApplication } from "@nestjs/common";
-import request from "supertest";
-import { AppModule } from "./../src/app.module";
-import { PrismaService } from "../src/prisma/prisma.service";
-import { ApiDocumentationService } from "../src/shared/api-documentation.service";
-
-describe("ProductsController (e2e)", () => {
-  let app: INestApplication;
-  let prisma: PrismaService;
-
-  const fixture = {
-    productId: "",
-    productDbId: "",
-    relatedProductDbId: "",
-    categoryId: "",
-    storeBrandId: "",
-    storeIds: [] as string[],
-  };
-  const isoDateRegex = /^\d{4}-\d{2}-\d{2}T\d{2}:\d{2}:\d{2}\.\d{3}Z$/;
-
-  const expectIsoDate = (value: unknown) => {
-    expect(typeof value).toBe("string");
-    expect(String(value)).toMatch(isoDateRegex);
-  };
-
-  const expectOfferContract = (
-    offer: Record<string, unknown>,
-    expectedBrandPrefix: string,
-  ) => {
-    expect(offer).toEqual(
-      expect.objectContaining({
-        id: expect.any(String),
-        currentPrice: expect.any(Number),
-        discountPrice: expect.anything(),
-        effectivePrice: expect.any(Number),
-        oldPrice: expect.any(Number),
-        discountPercent: expect.anything(),
-        availability: "in_stock",
-        updatedAt: expect.any(String),
-      }),
-    );
-
-    expect(offer.store).toEqual(expect.any(Object));
-    const store = offer.store as Record<string, unknown>;
-    expect(store).toEqual(
-      expect.objectContaining({
-        id: expect.any(String),
-        brand: expect.any(String),
-        city: expect.any(String),
-        address: expect.any(String),
-      }),
-    );
-    expect(String(store.brand)).toContain(expectedBrandPrefix);
-    expect(offer.effectivePrice).toBeLessThanOrEqual(
-      Number(offer.currentPrice),
-    );
-    expectIsoDate(offer.updatedAt);
-  };
-
-  beforeAll(async () => {
-    const moduleFixture: TestingModule = await Test.createTestingModule({
-      imports: [AppModule],
-    }).compile();
-
-    app = moduleFixture.createNestApplication();
-    ApiDocumentationService.configure(app);
-
-    await app.init();
-
-    prisma = app.get(PrismaService);
-
-    const suffix = Date.now();
-    const category = await prisma.productCategory.create({
-      data: {
-        name: `E2E Category ${suffix}`,
-      },
-    });
-
-    const storeBrand = await prisma.storeBrand.create({
-      data: {
-        name: `E2E Store Brand ${suffix}`,
-      },
-    });
-
-    const [storeA, storeB] = await Promise.all([
-      prisma.localStore.create({
-        data: {
-          brandId: storeBrand.id,
-          city: "Kyiv",
-          address: `E2E address A ${suffix}`,
-          longitude: 30.5,
-          latitude: 50.45,
-          openingHour: "08:00",
-          closingHour: "22:00",
-        },
-      }),
-      prisma.localStore.create({
-        data: {
-          brandId: storeBrand.id,
-          city: "Lviv",
-          address: `E2E address B ${suffix}`,
-          longitude: 24.03,
-          latitude: 49.84,
-          openingHour: "08:00",
-          closingHour: "22:00",
-        },
-      }),
-    ]);
-
-    const product = await prisma.product.create({
-      data: {
-        productId: `E2E-PRODUCT-${suffix}`,
-        canonicalName: "E2E Main Product",
-        brand: "E2E Brand",
-        categoryId: category.id,
-        media: "https://example.com/e2e-main.jpg",
-        measurements: { weight: "500g" },
-        pricingLogic: { pricePer: "item" },
-      },
-    });
-
-    const relatedProduct = await prisma.product.create({
-      data: {
-        productId: `E2E-RELATED-${suffix}`,
-        canonicalName: "E2E Related Product",
-        brand: "E2E Brand",
-        categoryId: category.id,
-        media: "https://example.com/e2e-related.jpg",
-        measurements: { weight: "450g" },
-        pricingLogic: { pricePer: "item" },
-      },
-    });
-
-    const [mainOfferA, mainOfferB, relatedOffer] = await Promise.all([
-      prisma.offer.create({
-        data: {
-          storeId: storeA.id,
-          productId: product.id,
-          currentPrice: 100,
-          discountPrice: 90,
-        },
-      }),
-      prisma.offer.create({
-        data: {
-          storeId: storeB.id,
-          productId: product.id,
-          currentPrice: 95,
-        },
-      }),
-      prisma.offer.create({
-        data: {
-          storeId: storeA.id,
-          productId: relatedProduct.id,
-          currentPrice: 88,
-        },
-      }),
-    ]);
-
-    await prisma.priceHistory.createMany({
-      data: [
-        {
-          offerId: mainOfferA.id,
-          price: 98,
-          regularPrice: 105,
-          startDate: new Date(Date.now() - 7 * 24 * 60 * 60 * 1000),
-        },
-        {
-          offerId: mainOfferA.id,
-          price: 90,
-          regularPrice: 100,
-          startDate: new Date(),
-        },
-        {
-          offerId: mainOfferB.id,
-          price: 96,
-          regularPrice: 99,
-          startDate: new Date(Date.now() - 6 * 24 * 60 * 60 * 1000),
-        },
-        {
-          offerId: mainOfferB.id,
-          price: 95,
-          regularPrice: 97,
-          startDate: new Date(),
-        },
-        {
-          offerId: relatedOffer.id,
-          price: 88,
-          regularPrice: 92,
-          startDate: new Date(),
-        },
-      ],
-    });
-
-    fixture.productId = product.productId;
-    fixture.productDbId = product.id;
-    fixture.relatedProductDbId = relatedProduct.id;
-    fixture.categoryId = category.id;
-    fixture.storeBrandId = storeBrand.id;
-    fixture.storeIds = [storeA.id, storeB.id];
-  });
-
-  afterAll(async () => {
-    await prisma.product.deleteMany({
-      where: {
-        id: {
-          in: [fixture.productDbId, fixture.relatedProductDbId],
-        },
-      },
-    });
-
-    await prisma.localStore.deleteMany({
-      where: {
-        id: {
-          in: fixture.storeIds,
-        },
-      },
-    });
-
-    await prisma.storeBrand.deleteMany({
-      where: {
-        id: fixture.storeBrandId,
-      },
-    });
-
-    await prisma.productCategory.deleteMany({
-      where: {
-        id: fixture.categoryId,
-      },
-    });
-
-    await app.close();
-  });
-
-  describe("GET /api/v1/products/:id/card", () => {
-    it("returns 200 with product card", async () => {
-      const response = await request(app.getHttpServer())
-        .get(`/api/v1/products/${fixture.productId}/card`)
-        .expect(200);
-
-      expect(response.body.product).toEqual(
-        expect.objectContaining({
-          id: fixture.productDbId,
-          productId: fixture.productId,
-          canonicalName: "E2E Main Product",
-          brand: "E2E Brand",
-          category: expect.any(String),
-          media: expect.any(String),
-          measurements: expect.any(Object),
-          description: null,
-          calories: null,
-        }),
-      );
-      expect(response.body.pricingSummary).toEqual(
-        expect.objectContaining({
-          bestPrice: expect.any(Number),
-          currency: "UAH",
-        }),
-      );
-      expect(response.body.topOffers.length).toBeGreaterThan(0);
-      expectOfferContract(response.body.topOffers[0], "E2E Store Brand ");
-      expect(response.body.stats).toEqual(
-        expect.objectContaining({
-          minPrice30d: expect.any(Number),
-          maxPrice30d: expect.any(Number),
-          avgPrice30d: expect.any(Number),
-          priceTrend: expect.any(String),
-        }),
-      );
-      expect(response.body.badges).toEqual(expect.any(Array));
-      expect(response.body.availabilityStatus).toEqual(expect.any(String));
-      expect(response.body.userContext).toEqual(
-        expect.objectContaining({
-          favorite: expect.any(Boolean),
-          inComparison: expect.any(Boolean),
-          inCart: expect.any(Boolean),
-        }),
-      );
-      expect(response.body.meta).toEqual(
-        expect.objectContaining({
-          fetchedAt: expect.any(String),
-          cacheTtlSeconds: expect.any(Number),
-        }),
-      );
-      expectIsoDate(response.body.meta.fetchedAt);
-    });
-
-    it("returns 404 for unknown product", async () => {
-      const response = await request(app.getHttpServer())
-        .get("/api/v1/products/not-found-product/card")
-        .expect(404);
-
-      expect(String(response.body.message)).toContain("not found");
-    });
-  });
-
-  describe("GET /api/v1/products", () => {
-    it("returns 200 with a product catalog page", async () => {
-      const response = await request(app.getHttpServer())
-        .get("/api/v1/products")
-        .query({ page: 1, limit: 10, categoryId: fixture.categoryId })
-        .expect(200);
-
-      expect(response.body).toEqual(
-        expect.objectContaining({
-          total: expect.any(Number),
-          page: 1,
-          limit: 10,
-          totalPages: expect.any(Number),
-        }),
-      );
-      expect(Array.isArray(response.body.items)).toBe(true);
-      expect(response.body.items.length).toBeGreaterThan(0);
-      expect(response.body.items[0]).toEqual(
-        expect.objectContaining({
-          id: expect.any(String),
-          canonicalName: expect.any(String),
-          brand: expect.anything(),
-          categoryId: fixture.categoryId,
-          offers: expect.arrayContaining([
-            expect.objectContaining({
-              id: expect.any(String),
-              storeId: expect.any(String),
-              price: expect.any(Number),
-              regularPrice: expect.any(Number),
-              discountPercent: expect.anything(),
-            }),
-          ]),
-        }),
-      );
-    });
-
-    it("filters products by store and price range and returns offers for the selected filters", async () => {
-      const response = await request(app.getHttpServer())
-        .get("/api/v1/products")
-        .query({
-          page: 1,
-          limit: 10,
-          storeId: fixture.storeIds[1],
-          minPrice: 95,
-          maxPrice: 95,
-          sort: "price_asc",
-        })
-        .expect(200);
-
-      expect(response.body.items.length).toBeGreaterThan(0);
-
-      const items = response.body.items as Array<Record<string, unknown>>;
-      items.forEach((item) => {
-        expect(Array.isArray(item.offers)).toBe(true);
-
-        const offers = item.offers as Array<Record<string, unknown>>;
-        expect(offers.length).toBeGreaterThan(0);
-
-        offers.forEach((offer) => {
-          expect(offer.storeId).toBe(fixture.storeIds[1]);
-          expect(offer.price).toBe(95);
-          expect(offer.regularPrice).toBeGreaterThanOrEqual(95);
-        });
-      });
-    });
-
-    it("accepts extended sort values", async () => {
-      const response = await request(app.getHttpServer())
-        .get("/api/v1/products")
-        .query({ sort: "discount" })
-        .expect(200);
-
-      expect(response.body).toEqual(
-        expect.objectContaining({
-          page: 1,
-          limit: 20,
-        }),
-      );
-    });
-
-    it("returns 400 for invalid paging query", async () => {
-      const response = await request(app.getHttpServer())
-        .get("/api/v1/products")
-        .query({ page: 0 })
-        .expect(400);
-
-      const message = Array.isArray(response.body.message)
-        ? response.body.message.join(" ")
-        : String(response.body.message);
-      expect(message).toContain("page");
-    });
-  });
-
-  describe("GET /api/v1/products/categories", () => {
-    it("returns 200 with category tree", async () => {
-      const response = await request(app.getHttpServer())
-        .get("/api/v1/products/categories")
-        .expect(200);
-
-      expect(Array.isArray(response.body.categories)).toBe(true);
-      expect(response.body.categories.length).toBeGreaterThan(0);
-      expect(response.body.categories[0]).toEqual(
-        expect.objectContaining({
-          id: fixture.categoryId,
-          name: expect.any(String),
-          parentId: null,
-          productCount: expect.any(Number),
-          children: expect.any(Array),
-        }),
-      );
-    });
-
-    it("returns 404 for unknown parent category", async () => {
-      const response = await request(app.getHttpServer())
-        .get("/api/v1/products/categories")
-        .query({ parentId: "missing-category" })
-        .expect(404);
-
-      expect(String(response.body.message)).toContain("not found");
-    });
-  });
-
-  describe("GET /api/v1/products/:id/offers", () => {
-    it("returns 200 with sorted offers", async () => {
-      const response = await request(app.getHttpServer())
-        .get(`/api/v1/products/${fixture.productId}/offers`)
-        .query({ sort: "price" })
-        .expect(200);
-
-      expect(response.body).toEqual(
-        expect.objectContaining({
-          productId: fixture.productDbId,
-          total: expect.any(Number),
-        }),
-      );
-      expect(response.body.offers.length).toBeGreaterThan(0);
-      expectOfferContract(response.body.offers[0], "E2E Store Brand ");
-      expect(response.body.total).toBe(response.body.offers.length);
-
-      const [first, second] = response.body.offers;
-      if (first && second) {
-        expect(first.effectivePrice).toBeLessThanOrEqual(second.effectivePrice);
-      }
-    });
-
-    it("returns 404 for unknown product", async () => {
-      await request(app.getHttpServer())
-        .get("/api/v1/products/not-found-product/offers")
-        .expect(404);
-    });
-
-    it("returns 400 for invalid sort", async () => {
-      const response = await request(app.getHttpServer())
-        .get(`/api/v1/products/${fixture.productId}/offers`)
-        .query({ sort: "cheapest" })
-        .expect(400);
-
-      const message = Array.isArray(response.body.message)
-        ? response.body.message.join(" ")
-        : String(response.body.message);
-      expect(message).toContain("sort");
-    });
-
-    it("returns 400 for invalid inStock", async () => {
-      const response = await request(app.getHttpServer())
-        .get(`/api/v1/products/${fixture.productId}/offers`)
-        .query({ inStock: "yes" })
-        .expect(400);
-
-      const message = Array.isArray(response.body.message)
-        ? response.body.message.join(" ")
-        : String(response.body.message);
-      expect(message).toContain("inStock");
-    });
-  });
-
-  describe("GET /api/v1/products/:id/price-history", () => {
-    it("returns 200 with history points", async () => {
-      const response = await request(app.getHttpServer())
-        .get(`/api/v1/products/${fixture.productId}/price-history`)
-        .query({ period: "30d" })
-        .expect(200);
-
-      expect(response.body).toEqual(
-        expect.objectContaining({
-          productId: fixture.productDbId,
-          period: "30d",
-        }),
-      );
-      expect(Array.isArray(response.body.points)).toBe(true);
-      expect(response.body.points.length).toBeGreaterThan(0);
-      expect(response.body.points[0]).toEqual(
-        expect.objectContaining({
-          date: expect.any(String),
-          price: expect.any(Number),
-          regularPrice: expect.any(Number),
-          store: expect.objectContaining({
-            id: expect.any(String),
-            brand: expect.stringContaining("E2E Store Brand "),
-            city: expect.any(String),
-          }),
-        }),
-      );
-      expectIsoDate(response.body.points[0].date);
-      expect(response.body.stats).toEqual(
-        expect.objectContaining({
-          minPrice: expect.any(Number),
-          maxPrice: expect.any(Number),
-          avgPrice: expect.any(Number),
-          trend: expect.any(String),
-        }),
-      );
-    });
-
-    it("returns 404 for unknown product", async () => {
-      await request(app.getHttpServer())
-        .get("/api/v1/products/not-found-product/price-history")
-        .query({ period: "30d" })
-        .expect(404);
-    });
-
-    it("returns 400 for invalid period", async () => {
-      const response = await request(app.getHttpServer())
-        .get(`/api/v1/products/${fixture.productId}/price-history`)
-        .query({ period: "qwerty" })
-        .expect(400);
-
-      const message = Array.isArray(response.body.message)
-        ? response.body.message.join(" ")
-        : String(response.body.message);
-      expect(message).toContain("period");
-    });
-  });
-
-  describe("GET /api/v1/products/:id/related", () => {
-    it("returns 200 with related products", async () => {
-      const response = await request(app.getHttpServer())
-        .get(`/api/v1/products/${fixture.productId}/related`)
-        .query({ limit: 5 })
-        .expect(200);
-
-      expect(response.body).toEqual(
-        expect.objectContaining({
-          productId: fixture.productDbId,
-        }),
-      );
-      expect(Array.isArray(response.body.related)).toBe(true);
-      expect(response.body.related.length).toBeGreaterThan(0);
-      expect(response.body.related[0]).toEqual(
-        expect.objectContaining({
-          id: expect.any(String),
-          productId: expect.any(String),
-          canonicalName: expect.any(String),
-          brand: expect.anything(),
-          media: expect.any(String),
-          bestPrice: expect.anything(),
-          offersCount: expect.any(Number),
-        }),
-      );
-      expect(
-        response.body.related.some(
-          (item: { id: string }) => item.id === fixture.relatedProductDbId,
-        ),
-      ).toBe(true);
-    });
-
-    it("returns 404 for unknown product", async () => {
-      await request(app.getHttpServer())
-        .get("/api/v1/products/not-found-product/related")
-        .expect(404);
-    });
-
-    it("returns 400 when limit is below minimum", async () => {
-      const response = await request(app.getHttpServer())
-        .get(`/api/v1/products/${fixture.productId}/related`)
-        .query({ limit: 0 })
-        .expect(400);
-
-      const message = Array.isArray(response.body.message)
-        ? response.body.message.join(" ")
-        : String(response.body.message);
-      expect(message).toContain("limit");
-    });
-
-    it("returns 400 when limit is above maximum", async () => {
-      const response = await request(app.getHttpServer())
-        .get(`/api/v1/products/${fixture.productId}/related`)
-        .query({ limit: 21 })
-        .expect(400);
-
-      const message = Array.isArray(response.body.message)
-        ? response.body.message.join(" ")
-        : String(response.body.message);
-      expect(message).toContain("limit");
-    });
-
-    it("returns 400 when limit is not an integer", async () => {
-      const response = await request(app.getHttpServer())
-        .get(`/api/v1/products/${fixture.productId}/related`)
-        .query({ limit: "abc" })
-        .expect(400);
-
-      const message = Array.isArray(response.body.message)
-        ? response.body.message.join(" ")
-        : String(response.body.message);
-      expect(message).toContain("limit");
-    });
-  });
-
-  describe("GET /api/v1/stores", () => {
-    it("returns 200 with list of stores", async () => {
-      const response = await request(app.getHttpServer())
-        .get("/api/v1/stores")
-        .expect(200);
-
-      expect(response.body).toEqual(
-        expect.objectContaining({
-          stores: expect.any(Array),
-        }),
-      );
-      expect(response.body.stores.length).toBeGreaterThan(0);
-
-      const store = response.body.stores[0];
-      expect(store).toEqual(
-        expect.objectContaining({
-          id: expect.any(String),
-          brand: expect.any(String),
-          logo: expect.any([String, null]),
-          website: expect.any([String, null]),
-          locationCount: expect.any(Number),
-        }),
-      );
-    });
-
-    it("stores are sorted by brand name", async () => {
-      const response = await request(app.getHttpServer())
-        .get("/api/v1/stores")
-        .expect(200);
-
-      const stores = response.body.stores;
-      const brandNames = stores.map((store: { brand: string }) => store.brand);
-      const sortedNames = [...brandNames].sort();
-
-      expect(brandNames).toEqual(sortedNames);
-    });
-
-    it("locationCount matches store locations in database", async () => {
-      const response = await request(app.getHttpServer())
-        .get("/api/v1/stores")
-        .expect(200);
-
-      expect(response.body.stores.length).toBeGreaterThan(0);
-      const store = response.body.stores.find(
-        (s: { id: string }) => s.id === fixture.storeBrandId,
-      );
-      expect(store).toBeDefined();
-      expect(store.locationCount).toBeGreaterThan(0);
-    });
-
-    describe("GET /api/v1/stores/:storeId/products", () => {
-      it("returns 200 with store products", async () => {
-        const response = await request(app.getHttpServer())
-          .get(`/api/v1/stores/${fixture.storeIds[0]}/products`)
-          .expect(200);
-
-        expect(response.body).toEqual(
-          expect.objectContaining({
-            storeId: fixture.storeIds[0],
-            storeName: expect.any(String),
-            items: expect.any(Array),
-            total: expect.any(Number),
-            page: expect.any(Number),
-            limit: expect.any(Number),
-            totalPages: expect.any(Number),
-          }),
-        );
-      });
-
-      it("returns products with correct structure", async () => {
-        const response = await request(app.getHttpServer())
-          .get(`/api/v1/stores/${fixture.storeIds[0]}/products`)
-          .expect(200);
-
-        if (response.body.items.length > 0) {
-          const item = response.body.items[0];
-          expect(item).toEqual(
-            expect.objectContaining({
-              id: expect.any(String),
-              productId: expect.any(String),
-              canonicalName: expect.any(String),
-              brand: expect.anything(),
-              media: expect.any(String),
-              currentPrice: expect.any(Number),
-              regularPrice: expect.any(Number),
-              discountPercent: expect.anything(),
-              currency: "UAH",
-              availabilityStatus: "in_stock",
-            }),
-          );
-        }
-      });
-
-      it("returns 404 for unknown store", async () => {
-        const response = await request(app.getHttpServer())
-          .get("/api/v1/stores/unknown-store/products")
-          .expect(404);
-
-        expect(String(response.body.message)).toContain("not found");
-      });
-
-      it("filters by minDiscount", async () => {
-        const response = await request(app.getHttpServer())
-          .get(`/api/v1/stores/${fixture.storeIds[0]}/products`)
-          .query({ minDiscount: 50 })
-          .expect(200);
-
-        if (response.body.items.length > 0) {
-          response.body.items.forEach(
-            (item: { discountPercent: number | null }) => {
-              if (item.discountPercent !== null) {
-                expect(item.discountPercent).toBeGreaterThanOrEqual(50);
-              }
-            },
-          );
-        }
-      });
-
-      it("sorts by price ascending", async () => {
-        const response = await request(app.getHttpServer())
-          .get(`/api/v1/stores/${fixture.storeIds[0]}/products`)
-          .query({ sort: "price_asc" })
-          .expect(200);
-
-        const items = response.body.items;
-        if (items.length > 1) {
-          for (let i = 1; i < items.length; i++) {
-            expect(items[i].currentPrice).toBeGreaterThanOrEqual(
-              items[i - 1].currentPrice,
-            );
-          }
-        }
-      });
-
-      it("sorts by discount descending", async () => {
-        const response = await request(app.getHttpServer())
-          .get(`/api/v1/stores/${fixture.storeIds[0]}/products`)
-          .query({ sort: "discount" })
-          .expect(200);
-
-        const items = response.body.items;
-        if (items.length > 1) {
-          for (let i = 1; i < items.length; i++) {
-            const prevDiscount = items[i - 1].discountPercent ?? 0;
-            const currDiscount = items[i].discountPercent ?? 0;
-            expect(currDiscount).toBeLessThanOrEqual(prevDiscount);
-          }
-        }
-      });
-
-      it("respects page and limit parameters", async () => {
-        const response = await request(app.getHttpServer())
-          .get(`/api/v1/stores/${fixture.storeIds[0]}/products`)
-          .query({ page: 1, limit: 10 })
-          .expect(200);
-
-        expect(response.body.page).toBe(1);
-        expect(response.body.limit).toBe(10);
-        expect(response.body.items.length).toBeLessThanOrEqual(10);
-      });
-
-      it("returns 400 for invalid sort", async () => {
-        const response = await request(app.getHttpServer())
-          .get(`/api/v1/stores/${fixture.storeIds[0]}/products`)
-          .query({ sort: "invalid" })
-          .expect(400);
-
-        const message = Array.isArray(response.body.message)
-          ? response.body.message.join(" ")
-          : String(response.body.message);
-        expect(message).toContain("sort");
-      });
-
-      it("returns 400 for invalid minDiscount", async () => {
-        const response = await request(app.getHttpServer())
-          .get(`/api/v1/stores/${fixture.storeIds[0]}/products`)
-          .query({ minDiscount: 150 })
-          .expect(400);
-
-        const message = Array.isArray(response.body.message)
-          ? response.body.message.join(" ")
-          : String(response.body.message);
-        expect(message).toContain("minDiscount");
-      });
-    });
-  });
-
-  describe("API documentation", () => {
-    it("returns OpenAPI spec json", async () => {
-      const response = await request(app.getHttpServer())
-        .get("/api/docs-json")
-        .expect(200);
-
-      expect(response.body).toEqual(
-        expect.objectContaining({
-          openapi: expect.any(String),
-          info: expect.objectContaining({
-            title: "Shop Comparison Platform API",
-            version: "1.0",
-          }),
-        }),
-      );
-    });
-
-    it("documents products endpoints", async () => {
-      const response = await request(app.getHttpServer())
-        .get("/api/docs-json")
-        .expect(200);
-
-      expect(response.body.paths).toEqual(
-        expect.objectContaining({
-          "/api/v1/products": expect.any(Object),
-          "/api/v1/products/categories": expect.any(Object),
-          "/api/v1/products/{id}/card": expect.any(Object),
-          "/api/v1/products/{id}/offers": expect.any(Object),
-          "/api/v1/products/{id}/price-history": expect.any(Object),
-          "/api/v1/products/{id}/related": expect.any(Object),
-          "/api/v1/stores": expect.any(Object),
-          "/api/v1/stores/{storeId}/products": expect.any(Object),
-        }),
-      );
-    });
-
-    it("documents query parameters for offers, price-history and related", async () => {
-      const response = await request(app.getHttpServer())
-        .get("/api/docs-json")
-        .expect(200);
-
-      const productsParams =
-        response.body.paths["/api/v1/products"].get.parameters;
-      expect(productsParams).toEqual(
-        expect.arrayContaining([
-          expect.objectContaining({ name: "page", in: "query" }),
-          expect.objectContaining({ name: "limit", in: "query" }),
-          expect.objectContaining({ name: "search", in: "query" }),
-          expect.objectContaining({ name: "brand", in: "query" }),
-          expect.objectContaining({ name: "categoryId", in: "query" }),
-          expect.objectContaining({ name: "inStock", in: "query" }),
-          expect.objectContaining({ name: "sort", in: "query" }),
-        ]),
-      );
-
-      const categoriesParams =
-        response.body.paths["/api/v1/products/categories"].get.parameters;
-      expect(categoriesParams).toEqual(
-        expect.arrayContaining([
-          expect.objectContaining({ name: "parentId", in: "query" }),
-        ]),
-      );
-
-      const offersParams =
-        response.body.paths["/api/v1/products/{id}/offers"].get.parameters;
-      expect(offersParams).toEqual(
-        expect.arrayContaining([
-          expect.objectContaining({
-            name: "sort",
-            in: "query",
-          }),
-          expect.objectContaining({
-            name: "inStock",
-            in: "query",
-          }),
-        ]),
-      );
-
-      const periodParam = response.body.paths[
-        "/api/v1/products/{id}/price-history"
-      ].get.parameters.find(
-        (param: { name: string }) => param.name === "period",
-      );
-      expect(periodParam).toBeDefined();
-      expect(periodParam.schema).toEqual(
-        expect.objectContaining({
-          type: "string",
-          pattern: "^(\\d+)(d|w|m)$",
-          default: "30d",
-        }),
-      );
-
-      const limitParam = response.body.paths[
-        "/api/v1/products/{id}/related"
-      ].get.parameters.find(
-        (param: { name: string }) => param.name === "limit",
-      );
-      expect(limitParam).toBeDefined();
-      expect(limitParam.schema).toEqual(
-        expect.objectContaining({
-          type: "number",
-          minimum: 1,
-          maximum: 20,
-        }),
-      );
-
-      const storeProductsParams =
-        response.body.paths["/api/v1/stores/{storeId}/products"].get.parameters;
-      expect(storeProductsParams).toEqual(
-        expect.arrayContaining([
-          expect.objectContaining({ name: "page", in: "query" }),
-          expect.objectContaining({ name: "limit", in: "query" }),
-          expect.objectContaining({ name: "search", in: "query" }),
-          expect.objectContaining({ name: "categoryId", in: "query" }),
-          expect.objectContaining({ name: "minDiscount", in: "query" }),
-          expect.objectContaining({ name: "sort", in: "query" }),
-        ]),
-      );
-    });
-  });
-});
